% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-attach_annotation.R
\name{attach_annotation}
\alias{attach_annotation}
\title{add annotation to screen object}
\usage{
attach_annotation(scr, ann)
}
\arguments{
\item{scr}{a \code{data.frame} containing the screen object}

\item{ann}{a \code{data.frame} containing the library annotation or path to such}
}
\value{
a \code{data.frame}
}
\description{
Merge library annotation with screen object.
}
\details{
The library annotation is merged with the screen object,
dropping all plates not present in the screen.
Missing gene symbols are substituted with their respective well types.
}
