% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-check_scan.R
\name{check_scan}
\alias{check_scan}
\title{check scan completion}
\usage{
check_scan(path, output = "print")
}
\arguments{
\item{path}{path to directory where scans are stored}

\item{output}{type of output:
print/console for printing to the console,
return/list for returning a named list;
determined by partial match}
}
\description{
Tally image files after a scan andreport missing images.
}
\details{
The function firt lists all directories in a given location
and identifies recent scans by their numeric postfixes.
Thereafter, it lists all .tif files in the appropriate data directories
and prints the well names, for which the number of collected images
is different to the maximum of all wells.
}
