% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-drawmeabarplot.R
\name{drawmeabarplot}
\alias{drawmeabarplot}
\title{draw a screen overview as a barplot}
\usage{
drawmeabarplot(data, highlight, treshold = 2)
}
\arguments{
\item{data}{a \code{data.frame}; must contain the column "mean_zscore_gf"}

\item{highlight}{regular expression (case insensitive) that matches the name of a column
that contains a logical flag for the group of genes to highlight}

\item{treshold}{integer; z* score treshold used for hit selection; location of horizontal lines}
}
\value{
a ggplot object
}
\description{
Produces a barplot of mean z scores of the green fraction of sample wells, from lowest to highest.
The function is designed to highlight a group of genes (e.g. components of a complex)
but can be made to work with single genes.
}
\details{
Mean green fraction z scores of all sample wells will be drawn as grey bars.
A set of genes specified by the \code{highlight} argument will be highlighted
in \code{orangered4} (the bars will be slightly widened for clarity).
Highlighting is specified by a logical column that flags the proper genes.
This column must be put in \code{data} beforehand.
}
