% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-extend_label.R
\name{extend_label}
\alias{extend_label}
\title{add items to hover label}
\usage{
extend_label(data, label, factors)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{label}{character vector of hover labels to be extended}

\item{factors}{names of columns of \code{data}
to add to \code{label}; character vector}
}
\value{
A character vector of extended hover labels.
}
\description{
Recursively extend highlight text by adding more factors.
}
\details{
This function takes a character vector \code{label} and pastes
each of the columns of \code{data} specified by \code{factors}.
It works recursively, so if \code{factors} is an empty vector,
it will returned the unchanged \code{label}.
}
