% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-fetch_files.R
\name{fetch_files}
\alias{fetch_files}
\title{fetch ScanR result/export files}
\usage{
fetch_files(where.from, where.to, object = "Main", verbose = TRUE)
}
\arguments{
\item{where.from}{directory where all screen data is stored;
defaults to current working directory,
otherwise must be an existing directory}

\item{where.to}{directory where all files will be copied;
defaults to current working directory,
otherwise may be any directory;
a non-existing one will be created}

\item{object}{regular expression that defines which ParameterData files to copy;
defaults to Main; set to NULL to skip ParameterData files}

\item{verbose}{logical flag; specifies whether to report progress}
}
\description{
Copy all ScanR result files, chosen parameter data files, and the screen log
into a single location.
}
\details{
The master directory specified by \code{where.from} will contain scan directories,
which may contain exported result files stored in "Population Data" subdirectories.
They may also contain parameter data files, which are named
"ParameterData_\code{object}.txt".
All result files and parameter data files where object name matches \code{object}
will be copied to the \code{where.to} directory. Hidden files will be omitted.

Names of scan directories, which are usually plate numbers,
are added to all respective file names.

If the master directory contains files whose names match regular expressions
"screenlog" and "layout", they will also be copied.
}
