% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-flag_hits.R
\name{flag_hits}
\alias{flag_hits}
\alias{flag_hits.numeric}
\alias{flag_hits.data.frame}
\alias{flag_hits.grouped_df}
\title{flag hits}
\usage{
flag_hits(x, ...)

\method{flag_hits}{numeric}(x, stringency, ...)

\method{flag_hits}{data.frame}(x, stringency, vars, ...)

\method{flag_hits}{grouped_df}(x, stringency, vars, ...)
}
\arguments{
\item{x}{numeric vector or data frame}

\item{...}{arguments passed to methods}

\item{stringency}{numeric vector of length 1 or the same as \code{vars};
if 1 or larger, the summary criterion will be used;
if lower that 1, the fractional criterion will be used,}

\item{vars}{variables to flag}
}
\value{
For vector method, a named list containing the aggregated hitscore and the hit status.
For data frame method, a copy of \code{x} with added columns.
}
\description{
Flag well as hit, depending on stringency.
}
\details{
If the screen involves multiple replicates, hitscores should be aggregated
and compared to a stringency threshold to determine whether a well is a hit.
Stringency can be applied summarily (add all hit scores) or fractionally
(divide the absolute sum of hit scores by the number of replicates).
Each well receives a logical flag, marking the plate/well as a hit or not.

Thus, samples that receive positive as well as negative hitscores
may not be flagged during aggregation.

Note that for the fractional method missing values are not counted
towards the number of replicates.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: vector method

\item \code{data.frame}: data frame method \cr
Aggregates hitscores for each variable named in \code{vars}. \cr
Ads 2 columns to \code{x} for each element of \code{vars}:
one for the aggregate (numeric) and one for the hit status (logical).

\item \code{grouped_df}: see \code{\link[metamethods]{data.frame__to__grouped_df}}
}}

\section{Grouping}{

This function should be run on subsets of the screen object, i.e. either
on a grouped data frame or on a data table with the \code{by} argument,
such that replicates are aggregated over plate/well combinations.
}

\section{Grouped data frames (\code{dplyr} package)}{

The method for class \code{grouped_df} is home made because the native behavior
of grouped data frames does not readily support functions that return data frames.
}

