% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlights.R
\name{highlights}
\alias{highlights}
\alias{highlight_genes}
\alias{color_sections}
\title{highlight genes}
\usage{
highlight_genes(canvas, data, genes, x.variable, y.variable,
  color.points = "lightyellow", color.highlight = "purple",
  plot.title = "")

color_sections(x.treshold = c(-2, 2), y.treshold = c(-2, 2),
  outer.limit = 100, categories = c("non-hit", "negative interaction",
  "positive interaction", "doubtful viability"),
  category.layout = matrix(c(4, 4, 4, 2, 1, 3, 2, 1, 3), nrow = 3, ncol =
  3), colpal = c("steelblue1", "springgreen2", "indianred2", "grey85"))
}
\arguments{
\item{canvas}{ggplot object produced by \code{color_sections}}

\item{data}{a \code{data.frame}}

\item{genes}{a character vector of \strong{gene symbols} to highlight}

\item{x.variable, y.variable}{names of variables to plot, given as character strings}

\item{color.points}{color to use for non-highlighted genes}

\item{color.highlight}{color to use for highlighted genes}

\item{plot.title}{character string passed to \code{link[ggplot2]{ggtitle}}}

\item{x.treshold, y.treshold, outer.limit}{coordinates that determine section borders
x.treshold and y.treshold may be determined
as numeric vectors of length one or two}

\item{categories}{names for sections, in order to appear in the plto legend}

\item{category.layout}{matrix that maps categories to sections; numbers refer to categories}

\item{colpal}{colors to assign to sections, in the same order as categories}
}
\value{
a ggplot object

a ggplot object with \code{geom_polygon} applied
}
\description{
Highlight selected genes on a scatter plot of all screen data.
}
\details{
Given a screen object (a \code{data.frame}), produce a scatter plot of two variables.
The background is divided into sections with colors signifying the hit status of respective
plot regions. Selected points (genes) are plotted with a different color
and their gene symbols are added as text labels.

Only sample wells will be plotted.
}
\section{Functions}{
\itemize{
\item \code{highlight_genes}: creates scatter plot on colored background

\item \code{color_sections}: colors sections of the plot background according to hit class
}}

