% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-hitscore.R
\name{hitscore}
\alias{hitscore}
\title{assign hit scoes}
\usage{
hitscore(x, treshold = 2)
}
\arguments{
\item{x}{a numeric vector}

\item{treshold}{absolute treshold for hit status}
}
\value{
a vector of hit statuses represented as integers: -1, 0 or 1
}
\description{
Assign hit status to single wells in a screen based on their z scores.
}
\details{
A data point is considered a hit if its hitting variable (usually a z score)
exceeds a treshold. A point can have one of three hitscores:
\itemize{
\item{
1 (positive hit) if its value is equal to or higher than the treshold
}
\item{
-1 (negative hit) if its value is equal or lowerto  than the negative treshold
}
\item{
0 (non-hit) if its absolute value is lower than the treshold
}
}
}
