% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-layouts.R
\name{layouts}
\alias{layouts}
\title{collate layout file for a screen}
\usage{
layouts(..., file)
}
\arguments{
\item{...}{files to load and collate, given as one or more character vectors}

\item{file}{file to save the layout in}
}
\value{
if \code{file} is missing, the collated layout table, otherwise nothing
}
\description{
Take any number of layout files and create a single layout file.
There must be exactly one file per plate type but each file may contain
layouts for different plating dates and these may be different.
}
\details{
This is a utility function for neatly putting together many layouts.
Sometimes a layout changes over the course of the screen. In that case it is easy (if a little tedious)
to prepare a composite layout file with a separate column for every plating date.
This function will load all such files and convert them into a single file
that can be loaded with \code{\link{build_screen}}.
If your screen maintains the same layout throughout, runninng this function is redundant.

The output file is a tab delimited text file that contains a narrow format table with the columns:
\code{well}, \code{well_type}, \code{plate_type}, and possibly \code{plated}.

Input files \strong{must} follow these rules:

1. must be tab delimited text files with a header

2. must have a "well" column with integer well numbers

3. must have the plate type encoded either as a "plate_type" column
\strong{or} in the file name: as the last element before the extension, e.g. "layout_S14_test.txt

4. if there is a single layout, it must be in a column called "well_type"

5. if there is more than one layout, each must be placed in a separate column,
the name of which must by the plating date in yyyymmdd format (legal separators are allowed)

There should also have a "position" column or "row" and "column" columns but this is optional.
}
\examples{
L <- layouts('layout_S14_test.txt',
             'layout_S14_control.txt')
head(L)
table(L$plated, L$well_type, L$plate_type)

layouts('layout_S14_test.txt',
        'layout_S14_control.txt',
        file = 'temp_layout_S14.txt')
head(read.delim('temp_layout_S14.txt'))

}
