% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-mask_values.R
\name{mask_values}
\alias{mask_values}
\title{mask values}
\usage{
mask_values(scr, cols, preds)
}
\arguments{
\item{scr}{a \code{data.frame}}

\item{cols}{character vector of column names to be masked}

\item{preds}{logical predicate that will determine masking
or a list of such predicates of the same length as \code{cols}:
a single predicate will be applied to all \code{cols},
multiple predicates will be applied to the respective \code{cols};
predicates can be given as bare expression or character strings}
}
\value{
A modified \code{data.frame} where the masked values are set to \code{NA}.
}
\description{
Mask certain values in the screen object to be omitted in downstream analysis.
}
\details{
Selected values (observation-variable fields) can be replaced with NAs,
so that they are omitted from normalization, standardization, and hit scoring.
}
\examples{
Iris <- rbind(head(iris), tail(iris))
mask_values(Iris, "Sepal.Length", Species == "setosa")
mask_values(Iris, "Sepal.Length", "Species == \"setosa\"")
mask_values(Iris, c("Sepal.Length", "Sepal.Width"), Species == "setosa")
mask_values(Iris, c("Sepal.Length", "Sepal.Width"), list(Species == "setosa", Species == "virginica"))
}
