% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-meanG.R
\name{meanG}
\alias{meanG}
\title{geometric mean}
\usage{
meanG(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{logical flag indicating whether to drop missing values}
}
\value{
Single numeric representing the geometric mean.
}
\description{
Compute the geometric mean.
}
