% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-normalize.R
\name{normalize}
\alias{normalize}
\title{normalize screen data}
\usage{
normalize(scr, method)
}
\arguments{
\item{scr}{screen object, a data frame}

\item{method}{normalization method, see \code{Details}}
}
\value{
an invisible \code{data frame}
}
\description{
Normalize data by one the (currently three) available methods:

\code{mean}: subtract reference mean from each data point

\code{median}: subtract reference median from each data point

\code{Tukey}: run Tukey's median Polish
}
\details{
There will be some details here...
}
