% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-patch_annotation.R
\name{patch_annotation}
\alias{patch_annotation}
\title{fix missing gene types in annotation file}
\usage{
patch_annotation(x)
}
\arguments{
\item{x}{\code{data.frame}
or path to file readable by \code{\link[data.table]{fread}}}
}
\value{
The annotation table with all bad rows fixed.
}
\description{
Patch errors in annotation that escaped \code{\link[acutils]{retry}}.
}
\details{
Even though \code{update_annotation} uses \code{acutils::retry}
to alleviate random errors that occur when calling \code{reutils::efetch} in bulk,
some queries still return errors. These will manifest as a "failed to retrieve"
value in \code{gene_type}.

This function extracts the erroneous rows from a file or table,
saves them to a temporary file in the current working directory,
and runs \code{update_annotation} on that file.
The fixed rows replace the bad ones in the original table and the fixed talbe is returned.

Once called, will run recursively until no errors are left.
Will throw a warning if there are many errors to fix.
}
