% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-percentage.R
\name{percentage}
\alias{percentage}
\alias{percentage.data.frame}
\alias{percentage.grouped_df}
\title{normalized percent inhibition/activation}
\usage{
percentage(x, variables, positive, negative, mode = c("inhibition",
  "activation"))

\method{percentage}{data.frame}(x, variables, positive, negative,
  mode = c("inhibition", "activation"))

\method{percentage}{grouped_df}(x, variables, positive, negative,
  mode = c("inhibition", "activation"))
}
\arguments{
\item{x}{a data frame}

\item{variables}{column to process}

\item{positive}{logical predicate to define positive control observations
(max inhibition); bare or string}

\item{negative}{logical predicate to define negative control observations
(no inhibition); bare or string}

\item{mode}{character that decides whether to calculate
normalized percent inhibition or normalized percent activation}
}
\value{
a modified \code{data.frame}
}
\description{
Calculate normalized percent inhibition (NPI) or activation (NPA)
for variable(s) in a data frame.
}
\details{
This is a method of data normalization in a high throughput screening campaign.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: uses \code{positive} and \code{negative} to calculate mean reference values,
runs function for NPI or NPA over desired variables with \code{lapply},
then \code{cbind}s the result to \code{x}

\item \code{grouped_df}: see \code{\link[metamethods]{data.frame__to__grouped_df}}
}}

\section{Grouped data frames (\code{dplyr} package)}{

The method for class \code{grouped_df} is home brewed as I don't know
how to properly handle this class.
}

\section{Grouping variables}{

Grouped data frames (class \code{grouped_df}) may cause problems
if grouping variables are used to define positive and negative controls.
It is best to create a separate variable to base the
\code{positive} and \code{negative} predicates on.
}

