% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-plate.type.converter.R
\name{plate.type.converter}
\alias{plate.type.converter}
\title{reformat plate type and replica designation}
\usage{
plate.type.converter(x, key)
}
\arguments{
\item{x}{a screen object}

\item{key}{optional dictinoary file or object, see \code{Details}}
}
\value{
screen object in which columns "plate_type" and "replica" are altered
}
\description{
This is an internal function called by \code{build_screen}.
It checks the plate type encoded in field 4 of the plate name and changes the columns
\code{plate_type} and \code{replica}, according to a dictionary.
}
\details{
The dictionary is supplied as a data frame, by default one found in a file
in the package's \code{extdata} directory. The file contains five options as of May 2019
and more options can easily be added by the user.

Alternatively, the dictionary can be taken from a custom file or an object specified by
the \code{key} argument. Be careful to keep to the format of the default file.

To view or edit the dictionary file find it with \code{path.package('siscreenr')}.
}
\keyword{internal}
