% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-plot_screen_progress.R
\name{plot_screen_progress}
\alias{plot_screen_progress}
\title{plot campaign progress over time}
\usage{
plot_screen_progress(directory, file)
}
\arguments{
\item{directory}{analysis master directory, where screen log file is located;
defaults to current working directory}

\item{file}{optional path to print the plot to;
will be created in the "results" directory;
set to NULL to send plot to current graphics device}
}
\value{
Nothing. A plot is printed if \code{file} is missing,
otherwise a 600x800 px png file is created in \code{directory}.
}
\description{
Create a barplot that shows how many plates were created and imaged
over the course of the screening campaign.
}
\details{
This function will search \code{directory} for a file that matches the pattern
"screenlog". Having found the file it will read plating and imaging dates
and plot the numbers of plates corresponding to those dates.
A barplot will be created with \code{ggplot}, which can be saved to a png file.

A subdirectory called "results" will be created, if absent, to save the plot.
}
