% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-uniques.R
\name{uniques}
\alias{uniques}
\title{meaningful factors}
\usage{
uniques(x)
}
\arguments{
\item{x}{a \code{data.frame}}
}
\value{
A character vector of column names.
}
\description{
Check for categorical variables with more than one unique value.
}
