% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-zfactor.R
\name{zfactor}
\alias{zfactor}
\title{calculate Z factor}
\usage{
zfactor(x, types)
}
\arguments{
\item{x}{numeric vector}

\item{types}{grouping factor (e.g. well types) to be passed to \code{tapply}
values \code{strong} must include "positive" and "negative"}
}
\value{
A single numeric.
}
\description{
Calculate the Z factor (also known as Z'factor) to estimate assay quality.
}
\details{
Z factor is used to estimate measurement accuracy in relation to the
dynamic range of the assay.

Z factor of 1 is perfect.  \cr
0.5 - 1 is a mean a good assay. \cr
0 - 0.5 means a weak assay. \cr
Z factor below 0 means the assay cannot be trusted.
This can be justifiable in some cases.
}
