% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-zscore.R
\name{zscore}
\alias{zscore}
\alias{zscore.default}
\alias{zscore.data.frame}
\alias{zscore.grouped_df}
\title{calculate (robust) zscore}
\usage{
zscore(x, ...)

\method{zscore}{default}(x, robust = TRUE, deviations = FALSE, reference, ...)

\method{zscore}{data.frame}(x, robust = TRUE, deviations = FALSE, reference, variables, ...)

\method{zscore}{grouped_df}(x, robust = TRUE, deviations = FALSE, reference, variables, ...)
}
\arguments{
\item{x}{a numeric vector, a data.frame, or a grouped data frame}

\item{...}{arguments passed to methods}

\item{robust}{logical flag; if TRUE (the default), calculate robust z scores}

\item{deviations}{logical flag;
set to FALSE (the default) if the supplied data is raw and
to TRUE (the default) if it has been normalized}

\item{reference}{optional specification of reference observations:
for the default method, a logical vector,
for the data frame method, either a logical vector
or any predicate (bare or as string) that refers to \code{x}'s variables}

\item{variables}{for data frame method, character vector of variables to standardize}
}
\value{
A numeric vector of z scores or a data frame with added columns of z scores.
}
\description{
Given a numeric vector, convert the numbers to z scores or robust z scores.
The function accepts raw values as well as normalized ones.
}
\details{
The z score is a data point standardized to the distribution it comes from.
It is a measure of the deviation of the point from the distribution's
location parameter expressed in terms of the dispersion parameter of that distribution.

For a normal z score the difference between a point and the distribution mean
is divided by the distribution's standard deviation. For a robust z score
the mean and standard deviation are replaced by the median and median absolute deviation,
respectively.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: for numeric vectors

\item \code{data.frame}: calculates zscores for requested variables in a data frame

\item \code{grouped_df}: see \code{\link[metamethods]{data.frame__to__grouped_df}}
}}

\section{Deviations}{

Data can be raw or normalized and this is specified with a logical flag.
For deviations the location parameter is assumed to be 0.
If normalization is done by simply subtracting the mean/median of the whole sample
from each data point, this is redundant but other normalization methods
may behave differently.
}

\section{Reference}{

Data points in the sample can be standardized against part
of the distribution rather than the whole. Should this be the case,
supply a logical vector (or predicate that will be evaluated
within the data frame) to determine the reference subpopulation.
}

\section{Grouped data frames (\code{dplyr} package)}{

The method for class \code{grouped_df} is home made because the native behavior
of grouped data frames does not readily support functions that return data frames.
}

