% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-attach_annotation.R
\name{attach_annotation}
\alias{attach_annotation}
\title{add annotation to screen object}
\usage{
attach_annotation(scr, ann)
}
\arguments{
\item{scr}{a \code{data.frame} containing the screen object}

\item{ann}{path to library annotation file;
if missing, defaults to internally stored annotation}
}
\value{
A \code{data.table}.
}
\description{
Merge library annotation with screen object.
}
\details{
The library annotation is merged with the screen object,
dropping all plates not present in the screen.
Missing gene symbols are substituted with their respective well types.
}
\section{Mismatching annotation}{

If the built-in annotation cannot be merged with the screen object,
an error will be thrown and the annotation will be assigned in the
Global Environment. It can then be modified and saved in a different file.
}

