% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-hitscore.R
\name{hitscore}
\alias{hitscore}
\title{assign hit scores}
\usage{
hitscore(x, threshold = 2)
}
\arguments{
\item{x}{a numeric vector}

\item{threshold}{numeric vector specifying the lower- and upper cut-off threshold;
normally of length 2, but a symmetric (e.g. -3 and 3) threshold
can be specified with a single numeric;}
}
\value{
a vector of hit statuses represented as integers: 1, -1 or 0
}
\description{
Assign hit status to single wells in a screen based on their z scores.
}
\details{
A data point is considered a hit if its hitting variable (usually a z score)
exceeds a threshold. A point can have one of three hitscores:
\itemize{
  \item{
  1 (positive hit) if its value is equal to or higher than the upper cut-off threshold
  }
  \item{
  -1 (negative hit) if its value is equal to or lower than the lower cut-off threshold
  }
  \item{
  0 (non-hit) otherwise
  }
}
}
