% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-normalize.R
\name{normalize}
\alias{normalize}
\alias{normalize.data.frame}
\alias{normalize.grouped_df}
\title{normalize screen data}
\usage{
normalize(x, variables, method = c("median", "mean", "medpolish"),
  reference)

\method{normalize}{data.frame}(x, variables, method = c("median", "mean",
  "medpolish"), reference)

\method{normalize}{grouped_df}(x, variables, method = c("median", "mean",
  "medpolish"), reference)
}
\arguments{
\item{x}{screen object, i.e. a \code{data.frame}, possibly \code{grouped}}

\item{variables}{variables to normalize;
character vector of column names or numeric vector of column indices}

\item{method}{normalization method, see \code{Details}}

\item{reference}{logical predicate that defines reference observations, bare or character}
}
\value{
an invisible \code{data.frame}
}
\description{
Normalize data, i.e. express each point as its deviation from a center.
}
\details{
There are three normalization methods available at the moment:
\itemize{
  \item{
  \code{mean}: subtract mean of reference from each data point
  }
  \item{
  \code{median}: subtract median of reference from each data point
  }
  \item{
  \code{medpolish}: run Tukey's median polish and return residuals;
                    calls \code{stats::medpolish}
  }
}
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: establishes normalization method, with possible \code{reference},
and runs it on desired variables with \code{lapply},
then \code{cbind}s the result to \code{x}

\item \code{grouped_df}: see \code{acutils::metamethod}
}}

\section{Warnings}{

If you are using the medpolish method, variables will be temporarily converted
from vectors to matrices. Make sure your data frames are ordered by column
(the default way matrices are filled) rather than by row (the default ScanR format).
Also, row and column specifications are necessary to read the matrix dimensions,
so \code{x} must contain either "row" and "column" variables or a "position" variable.

For other methods a reference subset can be specified. Any logical predicate will do.
If no reference is declared, normalization will be done against the whole population.
}

\section{Grouped data frames (\code{dplyr} package)}{

The method for class \code{grouped_df} is home brewed as I don't know
how to properly handle this class.
}

