% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-normalize.R
\name{normalize}
\alias{normalize}
\title{normalize screen data}
\usage{
normalize(scr, variables, reference, method = c("median", "mean",
  "medpolish"))
}
\arguments{
\item{scr}{screen object, a data frame}

\item{variables}{variables to normalize;
character vector of column names or numeric vector of column indices}

\item{reference}{logical predicate that defines reference observations}

\item{method}{normalization method, see \code{Details}}
}
\value{
an invisible \code{data.frame}
}
\description{
Normalize data, i.e. express each point as its deviation from a center.
}
\details{
There are three normalization methods available at the moment:
\itemize{
  \item{
  \code{mean}: subtract reference mean from each data point
  }
  \item{
  \code{median}: subtract reference median from each data point
  }
  \item{
  \code{medpolish}: run Tukey's median polish and return residuals;
                    calls \code{stats::medpolish}
  }
}
}
\section{Warnings}{

If you are using the medpolish method, variables will be temporarily converted
from vectors to matrices. Make sure your data frames are ordered by column
(the default way matrices are filled) rather than by row (the default ScanR format).

For other methods a reference subset can be specified.
This is done by filtering rows based on the \code{well_type} column, so there must be one.
If no reference is declared, normalization will be done against the whole population.
}

