% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-screenr.R
\docType{package}
\name{siscreenr}
\alias{siscreenr}
\alias{siscreenr-package}
\title{siscreenr: High Throughput Data Analysis}
\description{
A package for analyzing high throughput microscopy data produced by ScanR systems.
Provides tools for loading, normalizing and scoring data, and hit selection.
It is intended for interactive use so most of the functions leave a considerable amount of work to teh user.
}
\details{
The data is obtained with an automated fluorescence microscope and
preliminary image analysis is done with ScanR software.
Most experiments are done in 384 well plate format and usually involve dozens of plates.
However, it is compatible with other plate formats and some concepts can be transferred to slides as well.
}
\section{Functions}{

\itemize{
  \item{
  \code{build_screen} collate all data into a single data frame, reaarrange and add layout
  }
  \item{
  \code{normalize} normalize data by one of the available methods:
  subtract mean/median of reference or Tukey's median polish (for matrices)
  }
  \item{
  \code{zscore} calculate (robust) z scores
  }
  \item{
  \code{hitscore} classify data point as hit or not based on a given treshold; usually used on z scores
  }
  \item{
  \code{sumscore} calculate summary hit score and assign hit status to data point
  }
  \item{
  \code{fracscore} calculate fractional hit score and assign hit status to data point
  }
}
}

\author{
Aleksander Chlebowski
}
