% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-update_annotation.R
\name{update_annotation}
\alias{update_annotation}
\alias{check_geneid_status}
\alias{check_geneids}
\alias{get_gene_fields}
\alias{get_gene_fields_batch}
\alias{get_gene_type}
\alias{extract_field}
\title{update library annotation}
\usage{
update_annotation(infile, outfile, verbose = FALSE)

check_geneid_status(geneID)

check_geneids(geneIDs)

get_gene_fields(geneIDs)

get_gene_fields_batch(geneIDs)

get_gene_type(geneID)

extract_field(x)
}
\arguments{
\item{infile}{file containing the original annotation;
must be compatible with \code{\link[data.table]{fread}};
deafults to internally sotred Dharmacon annotation from 16th May 2015
(plate numbers have been unified, originally each subset was numbered independently)}

\item{outfile}{(optional) path to a file to save the updated annotation}

\item{verbose}{logical whether or not to report function progres}
}
\value{
The function either invisibly returns a data frame
        or saves to a specified path and returns nothing.
}
\description{
Check all geneIDs in library annotation file against GeneBank and get up-to-date information.
}
\details{
Since information in data bases can change it is prudent
to refresh the library annotation from time to time.
This function takes every geneID in an annotation file and checks its current status
in GeneBank: whether is has been withdrawn or replaced (and if so, by what new geneID).
It then queries Genebank again to retrieve the gene type (protein-coding, pseudogene, etc.).
Once the geneIDs are updated, yet another query is sent to GeneBank
to retrieve the current gene symbol, gene description, map location,
chromosome number, and aliases. This is done in batches of up to 499 items at a time.

Old geneIDs are kept in a separate column, \code{old_geneid}.
}
\section{Functions}{
\itemize{
\item \code{check_geneid_status}: checks a single geneID and retrieves its withdrawn and replaced status (TRUE/FALSE)
and a potential new geneID; then retrieves the gene type (protein-coding, pseudo, etc.);
there is a 1 second pause between queries; returns a character vector

\item \code{check_geneids}: runs \code{check_geneid_status} for all geneIDs and returns a \code{data.frame};
pauses for 0.5 second before each request to avoid swamping the server

\item \code{get_gene_fields}: queries the gene data base and retrieves five fields:
gene symbol, gene description, map location, chromosome number, and aliases
(other geneIDs associated with the geneID);

\item \code{get_gene_fields_batch}: runs \code{get_gene_fields} in batches of 499 and less; this is necessary as
the results of \code{reutils::efetch} are unworkable for larger sets

\item \code{get_gene_type}: queries the gene data base and retrieves the gene type field

\item \code{extract_field}: called by \code{get_gene_type} to extract the gene type
with provisions in case the field does not exist or is NA
}}

\section{File format}{

As of version 2.4 the function has undergone some generalization.
It now serves not only the original Dharmacon file but also other text files.
The input file may be tab- or comma delimited.
It must contain the following information:
plate number, well/position (e.g. A01), and geneID.
All other columns are immaterial but will not be dropped.

Annotation files can be updated again. In that case the columns \code{old_geneid}
and \code{old_gene_symbol} will remain as is and the update will be run with old geneIDs
rather than the updated ones.
}

\section{Dependencies}{

GeneBank queries are handled with the package \code{reutils}.
Data is loaded (and saved) with package \code{data.table}.
Data processing is done with (minimal) use of \code{dplyr} and \code{tidyr}.
Some errors are handled with \code{\link[retry]{acutils::retry}}.
Passing arguments is done with \code{\link[retry]{acutils::get.stack}}.
Several internal functions are called here, see \code{Functions}.
}

\section{Processing time}{

\code{check_geneid_status}
queries GeneBank one geneID at a time, which may swamp the server.
Hence, a 0.5 second pause is introduced before every query.
}

\keyword{internal}
